Database/database/requirements.txt
kagglehub>=0.3.3
pymongo>=4.6.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
